+++ client/src/components/QuoteFooter.vue
